requirements.txt
dotenv
streamlit
gitingest
google-genai